Separate out collision checking and collecting sensor data
Slightly cleaner overall, and opens up for future unit tests better.
