+++ include/move_basic/obstacle_points.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, Ubiquity Robotics
+ * Copyright (c) 2020, Ubiquity Robotics
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,13 +29,19 @@
  *
  */
 
-#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
-#include <tf2_ros/buffer.h>
-#include <sensor_msgs/Range.h>
-#include <sensor_msgs/LaserScan.h>
+#ifndef OBSTACLE_POINTS_H
+#define OBSTACLE_POINTS_H
 
+#include <vector>
+#include <utility>
 #include <mutex>
 
+#include <ros/ros.h>
+#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
+#include <tf2_ros/buffer.h>
+#include <tf2_ros/transform_listener.h>
+#include <sensor_msgs/Range.h>
+#include <sensor_msgs/LaserScan.h>
 
 // a single sensor with current obstacles
 class RangeSensor
@@ -61,81 +67,67 @@ public:
     void update(float range, ros::Time stamp);
 };
 
-// Handle Range messages and computes distance to obstacles
-class ObstacleDetector
+class ObstaclePoints
 {
-   // Line in polar form
-   class PolarLine
-   {
-   public:
-       float radius;
-       float theta;
- 
-       PolarLine(float radius, float theta);
-   };
+private:
+  // Line in polar form
+  class PolarLine
+  {
+  public:
+     float radius;
+     float theta;
 
-   std::string baseFrame;
+     PolarLine(float radius, float theta);
+  };
 
-   std::map<std::string, RangeSensor> sensors;
-   ros::Subscriber sonar_sub;
-   ros::Subscriber scan_sub;
-   ros::Publisher line_pub;
-   tf2_ros::Buffer *tf_buffer;
-   int sensor_id;
-   // footprint
-   float robot_width;
-   float robot_front_length;
-   float robot_back_length;
+  std::mutex points_mutex;
+   
+  std::string baseFrame;
 
-   float robot_width_sq;
-   float robot_front_length_sq;
-   float robot_back_length_sq;
-   float front_diag, back_diag;
+  std::map<std::string, RangeSensor> sensors;
+  ros::Subscriber sonar_sub;
+  ros::Subscriber scan_sub;
+  tf2_ros::Buffer tf_buffer;
+  tf2_ros::TransformListener tf_listener;
+   
+  bool have_lidar;
+  tf2::Vector3 lidar_origin;
+  tf2::Vector3 lidar_normal;
+  std::vector<PolarLine> lidar_points;
+  ros::Time lidar_stamp;
 
-   float max_age;
-   float no_obstacle_dist;
-   std::vector<tf2::Vector3> points;
-   bool have_test_points;
-   std::mutex obstacle_mutex;
-
-   bool have_lidar;
-   tf2::Vector3 lidar_origin;
-   tf2::Vector3 lidar_normal;
-   std::vector<PolarLine> lidar_points;
-
-   void draw_line(const tf2::Vector3 &p1, const tf2::Vector3 &p2,
-                  float r, float g, float b, int id);
-   void clear_line(int id);
-
-   void get_points();
-   void get_lidar_points(std::vector<tf2::Vector3>& points);
-
-   void check_dist(float x, bool forward, float& min_dist) const;
-   void check_angle(float theta, float x, float y,
-                    bool left, float& min_dist) const;
-
-   float degrees(float radians) const;
+  // Manually added points, used for unit testing things that
+  // use ObstaclePoints without having to go through ROS messages
+  std::vector<tf2::Vector3> test_points;
 
 public:
-   ros::Time stamp;
+  ObstaclePoints(ros::NodeHandle& nh);
 
-   ObstacleDetector(ros::NodeHandle& nh, tf2_ros::Buffer *tf_buffer);
-   void range_callback(const sensor_msgs::Range::ConstPtr &msg);
-   void scan_callback(const sensor_msgs::LaserScan::ConstPtr &msg);
+  void range_callback(const sensor_msgs::Range::ConstPtr &msg);
+  void scan_callback(const sensor_msgs::LaserScan::ConstPtr &msg);
 
-   // return distance in meters to closest obstacle
-   float obstacle_dist(bool forward, float &left_dist, float &right_dist,
-                       tf2::Vector3 &fl, tf2::Vector3 &fr);
+  /*
+   * Returns a vector of all the points that were detected, filtered
+   * by the maximum age.
+   *
+   */
+  std::vector<tf2::Vector3> get_points(ros::Duration max_age);
+ 
+  /*
+   * Returns a vector of lines (expressed as a pair of 2 points).
+   * The lines are based on the end of the sonar cones, filtered
+   * by the the specified maximum age.
+   *
+   */
+  typedef std::pair<tf2::Vector3, tf2::Vector3> Line;
+  std::vector<Line> get_lines(ros::Duration max_age);
 
-   // return distance in radians to closest obstacle
-   float obstacle_angle(bool left);
+  // Used for unit testing things that use ObstaclePoints 
+  // without having to go through ROS messages
+  void add_test_point(tf2::Vector3 p);
+  void clear_test_points();
 
-   // for testing
-   void add_test_point(tf2::Vector3 p);
-
-   double min_side_dist;
-   double max_side_dist;
-
-   void clear_test_points();
 };
+
+#endif
